Remove quantum identifier copy in line functions
Separated energy level map to do this easier and created therefore
the new files and the new includes
