+++ src/linefunctions.cc
@@ -240,9 +240,10 @@ void Linefunctions::set_lorentz(
     const Numeric& magnetic_magnitude,
     const Numeric& F0_noshift,
     const LineShape::Output& x,
+    const AbsorptionLines& band,
+    const Index& line_ind,
     const ArrayOfRetrievalQuantity& derivatives_data,
     const ArrayOfIndex& derivatives_data_position,
-    const QuantumIdentifier& quantum_identity,
     const LineShape::Output& dT,
     const LineShape::Output& dVMR) {
   constexpr Complex cpi(0, Constant::pi);
@@ -274,18 +275,18 @@ void Linefunctions::set_lorentz(
         dF.col(iq).noalias() = -iz * dw;
       else if ((deriv == JacPropMatType::LineCenter or
                 is_pressure_broadening_DV(deriv)) and
-               deriv.QuantumIdentity().In(quantum_identity))
+                Absorption::id_in_line(band, deriv.QuantumIdentity(), line_ind))
         dF.col(iq).noalias() = iz * dw;
       else if (is_pressure_broadening_G0(deriv) and
-               deriv.QuantumIdentity().In(quantum_identity))
+        Absorption::id_in_line(band, deriv.QuantumIdentity(), line_ind))
         dF.col(iq).noalias() = dw;
       else if (is_pressure_broadening_D0(deriv) and
-               deriv.QuantumIdentity().In(quantum_identity))
+        Absorption::id_in_line(band, deriv.QuantumIdentity(), line_ind))
         dF.col(iq).noalias() = iz * dw;
       else if (is_magnetic_parameter(deriv))
         dF.col(iq).noalias() = iz * zeeman_df * dw;
       else if (deriv == JacPropMatType::VMR) {
-        if (deriv.QuantumIdentity().In(quantum_identity))
+        if (Absorption::id_in_line(band, deriv.QuantumIdentity(), line_ind))
           dF.col(iq).noalias() = Complex(dVMR.G0, dVMR.D0 + dVMR.DV) * dw;
         else
           dF.col(iq).setZero();
@@ -304,9 +305,10 @@ void Linefunctions::set_voigt(
     const Numeric& F0_noshift,
     const Numeric& GD_div_F0,
     const LineShape::Output& x,
+    const AbsorptionLines& band,
+    const Index& line_ind,
     const ArrayOfRetrievalQuantity& derivatives_data,
     const ArrayOfIndex& derivatives_data_position,
-    const QuantumIdentifier& quantum_identity,
     const Numeric& dGD_div_F0_dT,
     const LineShape::Output& dxdT,
     const LineShape::Output& dxdVMR) {
@@ -350,18 +352,18 @@ void Linefunctions::set_voigt(
             F * dGD_dT * invGD - dw.cwiseProduct(z) * dGD_dT * invGD;
       else if ((deriv == JacPropMatType::LineCenter or
                 is_pressure_broadening_DV(deriv)) and
-               deriv.QuantumIdentity().In(quantum_identity))
+               Absorption::id_in_line(band, deriv.QuantumIdentity(), line_ind))
         dF.col(iq).noalias() = -F / F0 - dw * invGD - dw.cwiseProduct(z) / F0;
       else if (is_pressure_broadening_G0(deriv) and
-               deriv.QuantumIdentity().In(quantum_identity))
+               Absorption::id_in_line(band, deriv.QuantumIdentity(), line_ind))
         dF.col(iq).noalias() = iz * dw * invGD;
       else if (is_pressure_broadening_D0(deriv) and
-               deriv.QuantumIdentity().In(quantum_identity))
+               Absorption::id_in_line(band, deriv.QuantumIdentity(), line_ind))
         dF.col(iq).noalias() = -dw * invGD;
       else if (is_magnetic_parameter(deriv))
         dF.col(iq).noalias() = dw * (-zeeman_df * invGD);
       else if (deriv == JacPropMatType::VMR and
-               deriv.QuantumIdentity().In(quantum_identity))
+               Absorption::id_in_line(band, deriv.QuantumIdentity(), line_ind))
         dF.col(iq).noalias() =
             dw * Complex(-dxdVMR.D0 - dxdVMR.DV, dxdVMR.G0) * invGD;
       else
@@ -379,9 +381,10 @@ void Linefunctions::set_doppler(
     const Numeric& magnetic_magnitude,
     const Numeric& F0_noshift,
     const Numeric& GD_div_F0,
+    const AbsorptionLines& band,
+    const Index& line_ind,
     const ArrayOfRetrievalQuantity& derivatives_data,
     const ArrayOfIndex& derivatives_data_position,
-    const QuantumIdentifier& quantum_identity,
     const Numeric& dGD_div_F0_dT) {
   auto nppd = derivatives_data_position.nelem();
 
@@ -406,7 +409,7 @@ void Linefunctions::set_doppler(
       dF.col(iq).noalias() =
           -dGD_div_F0_dT * F0 * invGD * (2.0 * F.cwiseProduct(mx2) + F);
     else if (deriv == JacPropMatType::LineCenter and
-             deriv.QuantumIdentity().In(quantum_identity))
+             Absorption::id_in_line(band, deriv.QuantumIdentity(), line_ind))
       dF.col(iq).noalias() = -F.cwiseProduct(mx2) * (2 / F0) +
                              F.cwiseProduct(x) * 2 * (invGD - 1 / F0);
     else if (deriv == JacPropMatType::VMR)
@@ -421,9 +424,10 @@ void Linefunctions::apply_linemixing_scaling_and_mirroring(
     const Eigen::Ref<Eigen::MatrixXcd> dFm,
     const LineShape::Output& X,
     const bool with_mirroring,
+    const AbsorptionLines& band,
+    const Index& line_ind,
     const ArrayOfRetrievalQuantity& derivatives_data,
     const ArrayOfIndex& derivatives_data_position,
-    const QuantumIdentifier& quantum_identity,
     const LineShape::Output& dT,
     const LineShape::Output& dVMR) {
   auto nppd = derivatives_data_position.nelem();
@@ -441,13 +445,13 @@ void Linefunctions::apply_linemixing_scaling_and_mirroring(
         const auto c = Complex(dT.G, -dT.Y);
         dF.col(iq).noalias() += F * c + Fm * conj(c);
       } else if (is_pressure_broadening_G(deriv) and
-                 deriv.QuantumIdentity().In(quantum_identity))
+                 Absorption::id_in_line(band, deriv.QuantumIdentity(), line_ind))
         dF.col(iq).noalias() = F + Fm;
       else if (is_pressure_broadening_Y(deriv) and
-               deriv.QuantumIdentity().In(quantum_identity))
+               Absorption::id_in_line(band, deriv.QuantumIdentity(), line_ind))
         dF.col(iq).noalias() = Complex(0, -1) * (F - Fm);
       else if (deriv == JacPropMatType::VMR and
-               deriv.QuantumIdentity().In(quantum_identity)) {
+               Absorption::id_in_line(band, deriv.QuantumIdentity(), line_ind)) {
         const auto c = Complex(dVMR.G, -dVMR.Y);
         dF.col(iq).noalias() += F * c + Fm * conj(c);
       }
@@ -459,13 +463,13 @@ void Linefunctions::apply_linemixing_scaling_and_mirroring(
       if (deriv == JacPropMatType::Temperature)
         dF.col(iq).noalias() += F * Complex(dT.G, -dT.Y);
       else if (is_pressure_broadening_G(deriv) and
-               deriv.QuantumIdentity().In(quantum_identity))
+               Absorption::id_in_line(band, deriv.QuantumIdentity(), line_ind))
         dF.col(iq).noalias() = F;
       else if (is_pressure_broadening_Y(deriv) and
-               deriv.QuantumIdentity().In(quantum_identity))
+               Absorption::id_in_line(band, deriv.QuantumIdentity(), line_ind))
         dF.col(iq).noalias() = Complex(0, -1) * F;
       else if (deriv == JacPropMatType::VMR and
-               deriv.QuantumIdentity().In(quantum_identity))
+               Absorption::id_in_line(band, deriv.QuantumIdentity(), line_ind))
         dF.col(iq).noalias() += F * Complex(dVMR.G, -dVMR.Y);
     }
   }
@@ -480,9 +484,10 @@ void Linefunctions::apply_rosenkranz_quadratic_scaling(
     const Eigen::Ref<const Eigen::VectorXd> f_grid,
     const Numeric& F0,
     const Numeric& T,
+    const AbsorptionLines& band,
+    const Index& line_ind,
     const ArrayOfRetrievalQuantity& derivatives_data,
-    const ArrayOfIndex& derivatives_data_position,
-    const QuantumIdentifier& quantum_identity) {
+    const ArrayOfIndex& derivatives_data_position) {
   auto nf = f_grid.size(), nppd = derivatives_data_position.nelem();
 
   const Numeric invF0 = 1.0 / F0;
@@ -511,7 +516,7 @@ void Linefunctions::apply_rosenkranz_quadratic_scaling(
       if (deriv == JacPropMatType::Temperature)
         dF(iv, iq) += dmafac_dT_div_fun * F[iv];
       else if (deriv == JacPropMatType::LineCenter and
-               deriv.QuantumIdentity().In(quantum_identity))
+               Absorption::id_in_line(band, deriv.QuantumIdentity(), line_ind))
         dF(iv, iq) +=
             (-invF0 - Constant::h / (2.0 * Constant::k * T *
                                      std::tanh(F0 * Constant::h /
@@ -530,9 +535,10 @@ void Linefunctions::apply_VVH_scaling(
     const Eigen::Ref<const Eigen::VectorXd> f_grid,
     const Numeric& F0,
     const Numeric& T,
+    const AbsorptionLines& band,
+    const Index& line_ind,
     const ArrayOfRetrievalQuantity& derivatives_data,
-    const ArrayOfIndex& derivatives_data_position,
-    const QuantumIdentifier& quantum_identity) {
+    const ArrayOfIndex& derivatives_data_position) {
   auto nppd = derivatives_data_position.nelem();
 
   // 2kT is constant for the loop
@@ -565,7 +571,7 @@ void Linefunctions::apply_VVH_scaling(
           F.cwiseQuotient(f_grid) +
           c1 * (F.cwiseQuotient(tanh_f) - F.cwiseProduct(tanh_f));
     else if (deriv == JacPropMatType::LineCenter and
-             deriv.QuantumIdentity().In(quantum_identity))
+             Absorption::id_in_line(band, deriv.QuantumIdentity(), line_ind))
       dF.col(iq).noalias() +=
           (-1.0 / F0 + c1 * tanh_f0part - c1 / tanh_f0part) * F;
   }
@@ -576,9 +582,10 @@ void Linefunctions::apply_VVW_scaling(
     Eigen::Ref<Eigen::MatrixXcd> dF,
     const Eigen::Ref<const Eigen::VectorXd> f_grid,
     const Numeric& F0,
+    const AbsorptionLines& band,
+    const Index& line_ind,
     const ArrayOfRetrievalQuantity& derivatives_data,
-    const ArrayOfIndex& derivatives_data_position,
-    const QuantumIdentifier& quantum_identity) {
+    const ArrayOfIndex& derivatives_data_position) {
   auto nf = f_grid.size(), nppd = derivatives_data_position.nelem();
 
   // denominator is constant for the loop
@@ -600,7 +607,7 @@ void Linefunctions::apply_VVW_scaling(
 
       // These partial derivatives are special
       if (deriv == JacPropMatType::LineCenter and
-          deriv.QuantumIdentity().In(quantum_identity))
+          Absorption::id_in_line(band, deriv.QuantumIdentity(), line_ind))
         dF(iv, iq) -= 2.0 * invF0 * F[iv];
       else if (is_frequency_parameter(deriv))
         dF(iv, iq) += 2.0 / f_grid[iv] * F[iv];
@@ -633,9 +640,10 @@ void Linefunctions::apply_linestrength_scaling_by_lte(
     const Numeric& isotopic_ratio,
     const Numeric& QT,
     const Numeric& QT0,
+    const AbsorptionLines& band,
+    const Index& line_ind,
     const ArrayOfRetrievalQuantity& derivatives_data,
     const ArrayOfIndex& derivatives_data_position,
-    const QuantumIdentifier& quantum_identity,
     const Numeric& dQT_dT) {
   auto nppd = derivatives_data_position.nelem();
 
@@ -658,10 +666,10 @@ void Linefunctions::apply_linestrength_scaling_by_lte(
                    K2 +
                    dboltzman_ratio_dT(K1, T, line.E0()) / K1 - invQT * dQT_dT);
     else if (deriv == JacPropMatType::LineStrength and
-             deriv.QuantumIdentity().In(quantum_identity))
+             Absorption::id_in_line(band, deriv.QuantumIdentity(), line_ind))
       dF.col(iq).noalias() = F / line.I0();  //nb. overwrite
     else if (deriv == JacPropMatType::LineCenter and
-             deriv.QuantumIdentity().In(quantum_identity))
+             Absorption::id_in_line(band, deriv.QuantumIdentity(), line_ind))
       dF.col(iq).noalias() +=
           F *
           dstimulated_relative_emission_dF0(gamma, gamma_ref, T, T0) /
@@ -688,9 +696,10 @@ void Linefunctions::apply_linestrength_scaling_by_vibrational_nlte(
     const Numeric& isotopic_ratio,
     const Numeric& QT,
     const Numeric& QT0,
+    const AbsorptionLines& band,
+    const Index& line_ind,
     const ArrayOfRetrievalQuantity& derivatives_data,
     const ArrayOfIndex& derivatives_data_position,
-    const QuantumIdentifier& quantum_identity,
     const Numeric& dQT_dT) {
   auto nppd = derivatives_data_position.nelem();
 
@@ -736,11 +745,11 @@ void Linefunctions::apply_linestrength_scaling_by_vibrational_nlte(
       dN.col(iq).noalias() += F * (dS_dT_src - dS_dT_abs);
       dF.col(iq).noalias() += F * dS_dT_abs;
     } else if (deriv == JacPropMatType::LineStrength and
-               deriv.QuantumIdentity().In(quantum_identity)) {
+               Absorption::id_in_line(band, deriv.QuantumIdentity(), line_ind)) {
       dF.col(iq).noalias() = F * dS_dS0_abs;
       dN.col(iq).noalias() = F * (dS_dS0_src - dS_dS0_abs);
     } else if (deriv == JacPropMatType::LineCenter and
-               deriv.QuantumIdentity().In(quantum_identity)) {
+               Absorption::id_in_line(band, deriv.QuantumIdentity(), line_ind)) {
       const Numeric dS_dF0_abs =
           S_abs *
           (dstimulated_relative_emission_dF0(gamma, gamma_ref, T, T0) /
@@ -754,13 +763,13 @@ void Linefunctions::apply_linestrength_scaling_by_vibrational_nlte(
       dN.col(iq).noalias() += F * (dS_dF0_src - dS_dF0_abs);
       dF.col(iq).noalias() += F * dS_dF0_abs;
     } else if (deriv == JacPropMatType::NLTE) {
-      if (deriv.QuantumIdentity().InLower(quantum_identity)) {
+      if (Absorption::id_in_line_lower(band, deriv.QuantumIdentity(), line_ind)) {
         const Numeric dS_dTl_abs =
             S_abs * dabsorption_nlte_rate_dTl(gamma, T, Tl, Evl, r_low) / K3;
 
         dN.col(iq).noalias() = -F * dS_dTl_abs;
         dF.col(iq).noalias() = -dN.col(iq);
-      } else if (deriv.QuantumIdentity().InUpper(quantum_identity)) {
+      } else if (Absorption::id_in_line_upper(band, deriv.QuantumIdentity(), line_ind)) {
         const Numeric dS_dTu_abs =
             S_abs * dabsorption_nlte_rate_dTu(gamma, T, Tu, Evu, K4) / K3;
         const Numeric dS_dTu_src = S_src * dboltzman_ratio_dT(K4, Tu, Evu) / K4;
@@ -775,96 +784,22 @@ void Linefunctions::apply_linestrength_scaling_by_vibrational_nlte(
   F *= S_abs;
 }
 
-void Linefunctions::apply_linestrength_from_full_linemixing(
-    Eigen::Ref<Eigen::VectorXcd> F,
-    Eigen::Ref<Eigen::MatrixXcd> dF,
-    const Numeric& F0,
-    const Numeric& T,
-    const Complex& S_LM,
-    const Numeric& isotopic_ratio,
-    const ArrayOfRetrievalQuantity& derivatives_data,
-    const ArrayOfIndex& derivatives_data_position,
-    const QuantumIdentifier& quantum_identity,
-    const Complex& dS_LM_dT) {
-  const Index nppd = derivatives_data_position.nelem();
-
-  constexpr Numeric C1 = -Constant::h / Constant::k;
-
-  const Numeric invT = 1.0 / T, F0_invT = F0 * invT,
-                exp_factor = std::exp(C1 * F0_invT),
-                f0_factor = F0 * (1.0 - exp_factor);
-
-  const Complex S = S_LM * f0_factor * isotopic_ratio;
-
-  for (Index iq = 0; iq < nppd; iq++) {
-    const auto& deriv = derivatives_data[derivatives_data_position[iq]];
-
-    if (deriv == JacPropMatType::Temperature) {
-      dF.col(iq) *= S_LM;
-      dF.col(iq).noalias() +=
-          F *
-          (dS_LM_dT * f0_factor + S_LM * C1 * F0_invT * F0_invT * exp_factor) *
-          isotopic_ratio;
-    } else if (deriv == JacPropMatType::LineStrength) {
-      if (deriv.QuantumIdentity().In(quantum_identity))
-        throw std::runtime_error("Not working yet");
-    } else if (deriv == JacPropMatType::LineCenter or
-               is_pressure_broadening_DV(deriv)) {
-      if (deriv.QuantumIdentity().In(quantum_identity))
-        throw std::runtime_error("Not working yet");
-    } else {
-      dF.col(iq) *= S;
-    }
-  }
-
-  F *= S;
-}
-
-void Linefunctions::apply_dipole(Eigen::Ref<Eigen::VectorXcd> F,
-                                 Eigen::Ref<Eigen::MatrixXcd> /*dF*/,
-                                 const Numeric& F0,
-                                 const Numeric& T,
-                                 const Numeric& d0,
-                                 const Numeric& rho,
-                                 const Numeric& isotopic_ratio,
-                                 const ArrayOfRetrievalQuantity&,
-                                 const ArrayOfIndex& derivatives_data_position,
-                                 const QuantumIdentifier&,
-                                 const Numeric&) {
-  // Output is d0^2 * rho * F * isotopic_ratio * F0 * (1-e^(hF0/kT))
-
-  constexpr Numeric C1 = -Constant::h / Constant::k;
-
-  const Index nppd = derivatives_data_position.nelem();
-
-  const Numeric S = d0 * d0 * rho * isotopic_ratio, invT = 1.0 / T,
-                F0_invT = F0 * invT, exp_factor = exp(C1 * F0_invT),
-                f0_factor = F0 * (1.0 - exp_factor);
-
-  if (nppd)
-    throw std::runtime_error(
-        "Cannot support Jacobian from dipole calculations yet");
-
-  F *= S * f0_factor;
-}
-
 void Linefunctions::apply_lineshapemodel_jacobian_scaling(
     Eigen::Ref<Eigen::MatrixXcd> dF,
+    const AbsorptionLines& band,
+    const Index& line_ind,
     const ArrayOfRetrievalQuantity& derivatives_data,
     const ArrayOfIndex& derivatives_data_position,
-    const QuantumIdentifier& quantum_identity,
-    const AbsorptionLines& band,
     const Numeric& T,
     const Numeric& P,
-    const Vector& vmrs,
-    size_t line_ind) {
+    const Vector& vmrs) {
   auto nppd = derivatives_data_position.nelem();
 
   for (auto iq = 0; iq < nppd; iq++) {
     const RetrievalQuantity& rt =
         derivatives_data[derivatives_data_position[iq]];
     if (is_lineshape_parameter(rt) and
-        rt.QuantumIdentity().In(quantum_identity))
+        Absorption::id_in_line(band, rt.QuantumIdentity(), line_ind))
       dF.col(iq) *= band.ShapeParameter_dInternal(line_ind, T, P, vmrs, rt);
   }
 }
@@ -918,9 +853,10 @@ void Linefunctions::apply_linestrength_from_nlte_level_distributions(
     const Numeric& A21,
     const Numeric& F0,
     const Numeric& T,
+  const AbsorptionLines& band,
+  const Index& line_ind,
     const ArrayOfRetrievalQuantity& derivatives_data,
-    const ArrayOfIndex& derivatives_data_position,
-    const QuantumIdentifier& quantum_identity) {
+    const ArrayOfIndex& derivatives_data_position) {
   // Size of the problem
   auto nppd = derivatives_data_position.nelem();
 
@@ -964,7 +900,7 @@ void Linefunctions::apply_linestrength_from_nlte_level_distributions(
       dN.col(iq).noalias() +=
           F * e * Constant::h * F0 * exp_T / (c2 * Constant::k * T * T);
     else if (deriv == JacPropMatType::LineCenter and
-             deriv.QuantumIdentity().In(quantum_identity)) {
+             Absorption::id_in_line(band, deriv.QuantumIdentity(), line_ind)) {
       Numeric done_over_b_df0 =
           Constant::h * exp_T / (c2 * Constant::k * T) - 3.0 * b / F0;
       Numeric de_df0 = c1 * r2 * A21;
@@ -973,12 +909,12 @@ void Linefunctions::apply_linestrength_from_nlte_level_distributions(
       dN.col(iq).noalias() += F * (e * done_over_b_df0 + de_df0 / b - dk_df0);
       dF.col(iq).noalias() += F * dk_df0;
     } else if (deriv == JacPropMatType::NLTE) {
-      if (deriv.QuantumIdentity().InLower(quantum_identity)) {
+      if (Absorption::id_in_line_lower(band, deriv.QuantumIdentity(), line_ind)) {
         const Numeric dk_dr2 = -c3 * A21 / c2, de_dr2 = c3 * A21,
                       dratio_dr2 = de_dr2 / b - dk_dr2;
         dN.col(iq).noalias() = F * dratio_dr2;
         dF.col(iq).noalias() += F * dk_dr2;
-      } else if (deriv.QuantumIdentity().InUpper(quantum_identity)) {
+      } else if (Absorption::id_in_line_upper(band, deriv.QuantumIdentity(), line_ind)) {
         const Numeric dk_dr1 = c3 * x * A21 / c2;
         dF.col(iq).noalias() += F * dk_dr1;
       }
@@ -1000,9 +936,10 @@ void Linefunctions::set_htp(Eigen::Ref<Eigen::VectorXcd> F,
                             const Numeric& F0_noshift_si,
                             const Numeric& GD_div_F0_si,
                             const LineShape::Output& x_si,
+                            const AbsorptionLines& band,
+                            const Index& line_ind,
                             const ArrayOfRetrievalQuantity& derivatives_data,
                             const ArrayOfIndex& derivatives_data_position,
-                            const QuantumIdentifier& quantum_identity,
                             const Numeric& dGD_div_F0_dT_si,
                             const LineShape::Output& dxdT_si,
                             const LineShape::Output& dxdVMR_si) {
@@ -1127,39 +1064,39 @@ void Linefunctions::set_htp(Eigen::Ref<Eigen::VectorXcd> F,
                ((1 - x.ETA) * (dc0 - 1.5 * dc2));     // for T
         dc2t = (-c2 * dT.ETA) + ((1 - x.ETA) * dc2);  // for T
       } else if (rt == JacPropMatType::VMR and
-                 rt.QuantumIdentity().In(quantum_identity)) {
+                 Absorption::id_in_line(band, rt.QuantumIdentity(), line_ind)) {
         dc0 = Complex(dV.G0, -dV.D0);  // for VMR
         dc2 = Complex(dV.G2, -dV.D2);  // for VMR
         dc0t = (-(c0 - 1.5 * c2) * dV.ETA + dV.FVC) +
                ((1 - x.ETA) * (dc0 - 1.5 * dc2));     // for VMR
         dc2t = (-c2 * dV.ETA) + ((1 - x.ETA) * dc2);  // for VMR
       } else if (is_pressure_broadening_G2(rt) and
-                 rt.QuantumIdentity().In(quantum_identity)) {
+                 Absorption::id_in_line(band, rt.QuantumIdentity(), line_ind)) {
         dc2 =
             1;  // for Gam2(T, VMR) and for Shift2(T, VMR), the derivative is wrt c2 for later computations
         dc0t = ((1 - x.ETA) * (dc0 - 1.5 * dc2));
         dc2t = ((1 - x.ETA) * dc2);
       } else if (is_pressure_broadening_D2(rt) and
-                 rt.QuantumIdentity().In(quantum_identity)) {
+                 Absorption::id_in_line(band, rt.QuantumIdentity(), line_ind)) {
         dc2 =
             -iz;  // for Gam2(T, VMR) and for Shift2(T, VMR), the derivative is wrt c2 for later computations
         dc0t = ((1 - x.ETA) * (dc0 - 1.5 * dc2));
         dc2t = ((1 - x.ETA) * dc2);
       } else if (is_pressure_broadening_G0(rt) and
-                 rt.QuantumIdentity().In(quantum_identity)) {
+                 Absorption::id_in_line(band, rt.QuantumIdentity(), line_ind)) {
         dc0 =
             1;  // for Gam0(T, VMR) and for Shift0(T, VMR), the derivative is wrt c0 for later computations
         dc0t = ((1 - x.ETA) * (dc0 - 1.5 * dc2));
       } else if (is_pressure_broadening_D0(rt) and
-                 rt.QuantumIdentity().In(quantum_identity)) {
+                 Absorption::id_in_line(band, rt.QuantumIdentity(), line_ind)) {
         dc0 =
             -iz;  // for Gam0(T, VMR) and for Shift0(T, VMR), the derivative is wrt c0 for later computations
         dc0t = ((1 - x.ETA) * (dc0 - 1.5 * dc2));
       } else if (is_pressure_broadening_FVC(rt) and
-                 rt.QuantumIdentity().In(quantum_identity)) {
+                 Absorption::id_in_line(band, rt.QuantumIdentity(), line_ind)) {
         dc0t = (1);  // for FVC(T, VMR)
       } else if (is_pressure_broadening_ETA(rt) and
-                 rt.QuantumIdentity().In(quantum_identity)) {
+                 Absorption::id_in_line(band, rt.QuantumIdentity(), line_ind)) {
         dc0t = (-c0 + 1.5 * c2);  // for eta(T, VMR)
         dc2t = (-c2);             // for eta(T, VMR)
       }
@@ -1170,7 +1107,7 @@ void Linefunctions::set_htp(Eigen::Ref<Eigen::VectorXcd> F,
       if (is_magnetic_parameter(rt))
         dX = -iz * freq2kaycm(zeeman_df_si) / c2t;  // for H
       else if (rt == JacPropMatType::LineCenter and
-               rt.QuantumIdentity().In(quantum_identity))
+               Absorption::id_in_line(band, rt.QuantumIdentity(), line_ind))
         dX = -iz / c2t;  // for sg0
       else if (is_frequency_parameter(rt))
         dX = iz / c2t;  // for sg
@@ -1185,7 +1122,7 @@ void Linefunctions::set_htp(Eigen::Ref<Eigen::VectorXcd> F,
         if (is_magnetic_parameter(rt))
           dZ1 = -iz * cte * freq2kaycm(zeeman_df_si);  // for H
         else if (rt == JacPropMatType::LineCenter and
-                 rt.QuantumIdentity().In(quantum_identity))
+                 Absorption::id_in_line(band, rt.QuantumIdentity(), line_ind))
           dZ1 = -iz * cte;  // for sg0
         else if (is_frequency_parameter(rt))
           dZ1 = iz * cte;  // for sg
@@ -1213,7 +1150,7 @@ void Linefunctions::set_htp(Eigen::Ref<Eigen::VectorXcd> F,
         if (is_magnetic_parameter(rt))
           dZ1 = -iz * cte * freq2kaycm(zeeman_df_si);  // for H
         else if (rt == JacPropMatType::LineCenter and
-                 rt.QuantumIdentity().In(quantum_identity))
+                 Absorption::id_in_line(band, rt.QuantumIdentity(), line_ind))
           dZ1 = -iz * cte;  // for sg0
         else if (is_frequency_parameter(rt))
           dZ1 = iz * cte;  // for sg
@@ -1312,10 +1249,10 @@ void Linefunctions::set_htp(Eigen::Ref<Eigen::VectorXcd> F,
         dETA = dV.ETA;
         dFVC = dV.FVC;
       } else if (is_pressure_broadening_ETA(rt) and
-                 rt.QuantumIdentity().In(quantum_identity))
+                 Absorption::id_in_line(band, rt.QuantumIdentity(), line_ind))
         dETA = 1;
       else if (is_pressure_broadening_FVC(rt) and
-               rt.QuantumIdentity().In(quantum_identity))
+               Absorption::id_in_line(band, rt.QuantumIdentity(), line_ind))
         dFVC = 1;
 
       dF.col(iq)(iv) =
@@ -1399,6 +1336,12 @@ void Linefunctions::set_cross_section_of_band(
   Index start, nelem;
   find_cutoff_ranges(start, nelem, f_full, fcut_low, fcut_upp);
   
+  // VMR Jacobian
+  auto do_vmr = do_vmr_jacobian(derivatives_data, band.QuantumIdentity());
+  
+  // Placeholder nothingness
+  constexpr LineShape::Output empty_output = {0, 0, 0, 0, 0, 0, 0, 0, 0};
+  
   for (Index i=0; i<band.NumLines(); i++) {
     
     // Select the range of cutoff
@@ -1417,20 +1360,13 @@ void Linefunctions::set_cross_section_of_band(
     auto data = scratch.data.middleRows(start, nelem);
     const auto f = f_full.middleRows(start, nelem);
     
-    // Local line quantum information
-    const auto QI = band.QuantumIdentityOfLine(i);
-    
     // Pressure broadening and line mixing terms
     const auto X = band.ShapeParameters(i, T, P, vmrs);
-    
-    // Placeholder nothingness
-    constexpr LineShape::Output empty_output = {0, 0, 0, 0, 0, 0, 0, 0, 0};
     
     // Partial derivatives for temperature
     const auto dXdT = band.ShapeParameters_dT(i, T, P, vmrs);
     
     // Partial derivatives for VMR of self (function works for any species but only do self for now)
-    auto do_vmr = do_vmr_jacobian(derivatives_data, QI);
     const auto dXdVMR = do_vmr.test ?
     band.ShapeParameters_dVMR(i, T, P, do_vmr.qid) : empty_output;
     
@@ -1456,9 +1392,10 @@ void Linefunctions::set_cross_section_of_band(
                       H,
                       band.F0(i),
                       DC,
+                      band,
+                      i,
                       derivatives_data,
                       derivatives_data_active,
-                      QI,
                       dDCdT);
           if (band.Cutoff() not_eq Absorption::CutoffType::None)
             set_doppler(Fc,
@@ -1469,9 +1406,10 @@ void Linefunctions::set_cross_section_of_band(
                         H,
                         band.F0(i),
                         DC,
+                        band,
+                        i,
                         derivatives_data,
                         derivatives_data_active,
-                        QI,
                         dDCdT);
           break;
         case LineShape::Type::HTP:
@@ -1484,9 +1422,10 @@ void Linefunctions::set_cross_section_of_band(
                   band.F0(i),
                   DC,
                   X,
+                  band,
+                  i,
                   derivatives_data,
                   derivatives_data_active,
-                  QI,
                   dDCdT,
                   dXdT,
                   dXdVMR);
@@ -1499,9 +1438,10 @@ void Linefunctions::set_cross_section_of_band(
                     band.F0(i),
                     DC,
                     X,
+                    band,
+                    i,
                     derivatives_data,
                     derivatives_data_active,
-                    QI,
                     dDCdT,
                     dXdT,
                     dXdVMR);
@@ -1515,9 +1455,10 @@ void Linefunctions::set_cross_section_of_band(
                       H,
                       band.F0(i),
                       X,
+                      band,
+                      i,
                       derivatives_data,
                       derivatives_data_active,
-                      QI,
                       dXdT,
                       dXdVMR);
           if (band.Cutoff() not_eq Absorption::CutoffType::None)
@@ -1529,9 +1470,10 @@ void Linefunctions::set_cross_section_of_band(
                         H,
                         band.F0(i),
                         X,
+                        band,
+                        i,
                         derivatives_data,
                         derivatives_data_active,
-                        QI,
                         dXdT,
                         dXdVMR);
           break;
@@ -1545,9 +1487,10 @@ void Linefunctions::set_cross_section_of_band(
                     band.F0(i),
                     DC,
                     X,
+                    band,
+                    i,
                     derivatives_data,
                     derivatives_data_active,
-                    QI,
                     dDCdT,
                     dXdT,
                     dXdVMR);
@@ -1561,9 +1504,10 @@ void Linefunctions::set_cross_section_of_band(
                       band.F0(i),
                       DC,
                       X,
+                      band,
+                      i,
                       derivatives_data,
                       derivatives_data_active,
-                      QI,
                       dDCdT,
                       dXdT,
                       dXdVMR);
@@ -1596,9 +1540,10 @@ void Linefunctions::set_cross_section_of_band(
                       H,
                       -band.F0(i),
                       LineShape::mirroredOutput(X),
+                      band,
+                      i,
                       derivatives_data,
                       derivatives_data_active,
-                      QI,
                       LineShape::mirroredOutput(dXdT),
                       LineShape::mirroredOutput(dXdVMR));
           if (band.Cutoff() not_eq Absorption::CutoffType::None)
@@ -1610,9 +1555,10 @@ void Linefunctions::set_cross_section_of_band(
                         H,
                         -band.F0(i),
                         LineShape::mirroredOutput(X),
+                        band,
+                        i,
                         derivatives_data,
                         derivatives_data_active,
-                        QI,
                         LineShape::mirroredOutput(dXdT),
                         LineShape::mirroredOutput(dXdVMR));
           break;
@@ -1627,9 +1573,10 @@ void Linefunctions::set_cross_section_of_band(
                           H,
                           -band.F0(i),
                           -DC,
+                          band,
+                          i,
                           derivatives_data,
                           derivatives_data_active,
-                          QI,
                           -dDCdT);
               if (band.Cutoff() not_eq Absorption::CutoffType::None)
                 set_doppler(Nc,
@@ -1640,9 +1587,10 @@ void Linefunctions::set_cross_section_of_band(
                             H,
                             -band.F0(i),
                             -DC,
+                            band,
+                            i,
                             derivatives_data,
                             derivatives_data_active,
-                            QI,
                             -dDCdT);
               break;
             case LineShape::Type::LP:
@@ -1654,9 +1602,10 @@ void Linefunctions::set_cross_section_of_band(
                           H,
                           -band.F0(i),
                           LineShape::mirroredOutput(X),
+                          band,
+                          i,
                           derivatives_data,
                           derivatives_data_active,
-                          QI,
                           LineShape::mirroredOutput(dXdT),
                           LineShape::mirroredOutput(dXdVMR));
               if (band.Cutoff() not_eq Absorption::CutoffType::None)
@@ -1668,9 +1617,10 @@ void Linefunctions::set_cross_section_of_band(
                             H,
                             -band.F0(i),
                             LineShape::mirroredOutput(X),
+                            band,
+                            i,
                             derivatives_data,
                             derivatives_data_active,
-                            QI,
                             LineShape::mirroredOutput(dXdT),
                             LineShape::mirroredOutput(dXdVMR));
               break;
@@ -1684,9 +1634,10 @@ void Linefunctions::set_cross_section_of_band(
                         -band.F0(i),
                         -DC,
                         LineShape::mirroredOutput(X),
+                        band,
+                        i,
                         derivatives_data,
                         derivatives_data_active,
-                        QI,
                         -dDCdT,
                         LineShape::mirroredOutput(dXdT),
                         LineShape::mirroredOutput(dXdVMR));
@@ -1700,9 +1651,10 @@ void Linefunctions::set_cross_section_of_band(
                           -band.F0(i),
                           -DC,
                           LineShape::mirroredOutput(X),
+                          band,
+                          i,
                           derivatives_data,
                           derivatives_data_active,
-                          QI,
                           -dDCdT,
                           LineShape::mirroredOutput(dXdT),
                           LineShape::mirroredOutput(dXdVMR));
@@ -1718,9 +1670,10 @@ void Linefunctions::set_cross_section_of_band(
                       -band.F0(i),
                       -DC,
                       LineShape::mirroredOutput(X),
+                      band,
+                      i,
                       derivatives_data,
                       derivatives_data_active,
-                      QI,
                       -dDCdT,
                       LineShape::mirroredOutput(dXdT),
                       LineShape::mirroredOutput(dXdVMR));
@@ -1733,9 +1686,10 @@ void Linefunctions::set_cross_section_of_band(
                         -band.F0(i),
                         -DC,
                         LineShape::mirroredOutput(X),
+                        band,
+                        i,
                         derivatives_data,
                         derivatives_data_active,
-                        QI,
                         -dDCdT,
                         LineShape::mirroredOutput(dXdT),
                         LineShape::mirroredOutput(dXdVMR));
@@ -1752,27 +1706,27 @@ void Linefunctions::set_cross_section_of_band(
       // Mixing and mirroring can only apply to non-Doppler shapes
       if (band.LineShapeType() not_eq LineShape::Type::DP) {
         apply_linemixing_scaling_and_mirroring(F,
-                                              dF,
-                                              N,
-                                              dN,
-                                              X,
-                                              with_mirroring,
-                                              derivatives_data,
-                                              derivatives_data_active,
-                                              QI,
-                                              dXdT,
-                                              dXdVMR);
+                                               dF,
+                                               N,
+                                               dN,
+                                               X,
+                                               with_mirroring,
+                                               band,
+                                               i,
+                                               derivatives_data,
+                                               derivatives_data_active,
+                                               dXdT,
+                                               dXdVMR);
 
         // Apply line mixing and pressure broadening partial derivatives
         apply_lineshapemodel_jacobian_scaling(dF,
+                                              band,
+                                              i,
                                               derivatives_data,
                                               derivatives_data_active,
-                                              QI,
-                                              band,
                                               T,
                                               P,
-                                              vmrs,
-                                              i);
+                                              vmrs);
       }
 
       // Line normalization if necessary
@@ -1788,28 +1742,31 @@ void Linefunctions::set_cross_section_of_band(
                             f,
                             band.F0(i),
                             T,
+                            band,
+                            i,
                             derivatives_data,
-                            derivatives_data_active,
-                            QI);
+                            derivatives_data_active);
           break;
         case Absorption::NormalizationType::VVW:
           apply_VVW_scaling(F,
                             dF,
                             f,
                             band.F0(i),
+                            band,
+                            i,
                             derivatives_data,
-                            derivatives_data_active,
-                            QI);
+                            derivatives_data_active);
           break;
         case Absorption::NormalizationType::RosenkranzQuadratic:
           apply_rosenkranz_quadratic_scaling(F,
-                                            dF,
-                                            f,
-                                            band.F0(i),
-                                            T,
-                                            derivatives_data,
-                                            derivatives_data_active,
-                                            QI);
+                                             dF,
+                                             f,
+                                             band.F0(i),
+                                             T,
+                                             band,
+                                             i,
+                                             derivatives_data,
+                                             derivatives_data_active);
           break;
       }
 
@@ -1826,13 +1783,14 @@ void Linefunctions::set_cross_section_of_band(
                                             isot_ratio,
                                             QT,
                                             QT0,
+                                            band,
+                                            i,
                                             derivatives_data,
                                             derivatives_data_active,
-                                            QI,
                                             dQTdT);
           break;
         case Absorption::PopulationType::ByNLTEVibrationalTemperatures: {
-          auto nlte_data = nlte.get_vibtemp_params(QI, T);
+          auto nlte_data = nlte.get_vibtemp_params(band, i, T);
           apply_linestrength_scaling_by_vibrational_nlte(
               F,
               dF,
@@ -1848,13 +1806,14 @@ void Linefunctions::set_cross_section_of_band(
               isot_ratio,
               QT,
               QT0,
+              band,
+              i,
               derivatives_data,
               derivatives_data_active,
-              QI,
               dQTdT);
         } break;
         case Absorption::PopulationType::ByNLTEPopulationDistribution: {
-          auto nlte_data = nlte.get_ratio_params(QI);
+          auto nlte_data = nlte.get_ratio_params(band, i);
           apply_linestrength_from_nlte_level_distributions(
               F,
               dF,
@@ -1867,9 +1826,10 @@ void Linefunctions::set_cross_section_of_band(
               band.A(i),
               band.F0(i),
               T,
+              band,
+              i,
               derivatives_data,
-              derivatives_data_active,
-              QI);
+              derivatives_data_active);
         } break;
         case Absorption::PopulationType::ByRelmatMendazaLTE:
         case Absorption::PopulationType::ByRelmatHartmannLTE:
