+++ src/quantum.h
@@ -655,129 +655,6 @@ std::ostream& operator<<(std::ostream& os, const QuantumNumbers& qn);
 /** Output operator */
 std::ostream& operator<<(std::ostream& os, const QuantumIdentifier& qi);
 
-// Output from EnergyLevelMap
-struct Output2{
-  Numeric r_low;
-  Numeric r_upp;
-};
-
-// Output from EnergyLevelMap
-struct Output4{
-  Numeric E_low;
-  Numeric E_upp;
-  Numeric T_low;
-  Numeric T_upp;
-};
-
-enum class EnergyLevelMapType {
-  Tensor3_t,
-  Vector_t,
-  Numeric_t,
-  None_t,
-};
-
-EnergyLevelMapType string2energylevelmaptype(const String& s);
-
-String energylevelmaptype2string(EnergyLevelMapType type);
-
-class EnergyLevelMap {
-private:
-  EnergyLevelMapType mtype;
-  ArrayOfQuantumIdentifier mlevels;
-  Vector mvib_energy;
-  Tensor4 mvalue;
-  
-public:
-  void ThrowIfNotOK() const {
-    if (not (mvalue.nbooks() == mlevels.nelem() and 
-      (mvib_energy.nelem() == mlevels.nelem() or mvib_energy.nelem() == 0)))
-      throw std::runtime_error("Bad dimensions, vibrational energies and IDs and data of strange size");
-    if (mtype == EnergyLevelMapType::Tensor3_t) {
-    } else if (mtype == EnergyLevelMapType::Vector_t) {
-      if (mvalue.npages() not_eq 1 or mvalue.nrows() not_eq 1)
-        throw std::runtime_error("Bad dimensions for vector type");
-    } else if (mtype == EnergyLevelMapType::Numeric_t) {
-      if (mvalue.npages() not_eq 1 or mvalue.nrows() not_eq 1 or mvalue.ncols() not_eq 1)
-        throw std::runtime_error("Bad dimensions for numeric type");
-    } else if (mtype == EnergyLevelMapType::None_t) {
-      if (mvalue.npages() not_eq 0 or mvalue.nrows() not_eq 0 or mvalue.ncols() not_eq 0)
-        throw std::runtime_error("Bad dimensions for none type");
-    }
-    
-    for (auto& e: mvib_energy) if (e < 0) throw std::runtime_error("Bad energies");
-  }
-  
-  EnergyLevelMap() : mtype(EnergyLevelMapType::None_t), mlevels(0),
-  mvib_energy(0), mvalue(0, 0, 0, 0) {ThrowIfNotOK();}
-  
-  EnergyLevelMap(EnergyLevelMapType new_type, Index pages, Index rows,
-                 Index cols, const EnergyLevelMap& old) : 
-  mtype(new_type), mlevels(old.mlevels), mvib_energy(old.mvib_energy),
-  mvalue(old.mlevels.nelem(), pages, rows, cols) {ThrowIfNotOK();};
-  
-  // Create Tensor3_t from the raw inputs
-  EnergyLevelMap(const Tensor4& data, const ArrayOfQuantumIdentifier& levels, const Vector& energies=Vector(0));
-  
-  // Create Vector_t from the raw inputs
-  EnergyLevelMap(const Matrix& data, const ArrayOfQuantumIdentifier& levels, const Vector& energies=Vector(0));
-  
-  // Create Numeric_t from the raw inputs
-  EnergyLevelMap(const Vector& data, const ArrayOfQuantumIdentifier& levels, const Vector& energies=Vector(0));
-  
-  // Create Vector_t from Tensor3_t
-  EnergyLevelMap InterpToGridPos(Index atmosphere_dim, const ArrayOfGridPos& p, const ArrayOfGridPos& lat, const ArrayOfGridPos& lon) const;
-  
-  // Create Numeric_t from Vector_t
-  EnergyLevelMap operator[](Index ip) const;
-  
-  // Create Numeric_t from Tensor3_t
-  EnergyLevelMap operator()(Index ip, Index ilat, Index ilon) const;
-  
-  /** Energy level type */
-  EnergyLevelMapType Type() const noexcept {return mtype;}
-  
-  /** Energy level type */
-  const ArrayOfQuantumIdentifier& Levels() const noexcept {return mlevels;}
-  
-  /** Energy level type */
-  const Vector& Energies() const noexcept {return mvib_energy;}
-  
-  /** Energy level type */
-  const Tensor4& Data() const noexcept {return mvalue;}
-  
-  /** Energy level type */
-  EnergyLevelMapType& Type() noexcept {return mtype;}
-  
-  /** Energy level type */
-  ArrayOfQuantumIdentifier& Levels() noexcept {return mlevels;}
-  
-  /** Energy level type */
-  Vector& Energies() noexcept {return mvib_energy;}
-  
-  /** Energy level type */
-  Tensor4& Data() noexcept {return mvalue;}
-  
-  //////////////////////
-  // Numeric_t access //
-  //////////////////////
-  
-  /** Get the output required for Population::NLTE
-   * 
-   * @param[in] transition A line-by-line transition
-   * @return Upper and lower level distributions
-   */
-  Output2 get_ratio_params(const QuantumIdentifier& transition) const;
-  
-  /** Get the output required for Population::NLTE-VibrationalTemperatures
-   * 
-   * @param[in] transition A line-by-line transition
-   * @return Upper and lower level distributions and energies
-   */
-  Output4 get_vibtemp_params(const QuantumIdentifier& transition, const Numeric T) const;
-};
-
-std::ostream& operator<<(std::ostream& os, const EnergyLevelMap& elm);
-
 std::ostream& operator<<(std::ostream& os, QuantumNumberType t);
 
 #endif
